ISPN-6856 Test class-bloat reduction (15)
