+++ core/src/test/java/org/infinispan/tx/totalorder/statetransfer/TotalOrderStateTransferFunctionalTest.java
@@ -1,0 +1,65 @@
+package org.infinispan.tx.totalorder.statetransfer;
+
+import org.infinispan.configuration.cache.CacheMode;
+import org.infinispan.configuration.cache.VersioningScheme;
+import org.infinispan.statetransfer.StateTransferFunctionalTest;
+import org.infinispan.transaction.TransactionProtocol;
+import org.infinispan.util.concurrent.IsolationLevel;
+import org.testng.annotations.Factory;
+import org.testng.annotations.Test;
+
+/**
+ * @author Mircea Markus <mircea.markus@jboss.com> (C) 2011 Red Hat Inc.
+ * @since 5.3
+ */
+@Test(groups = "functional", testName = "tx.totalorder.statetransfer.TotalOrderStateTransferFunctionalTest")
+public class TotalOrderStateTransferFunctionalTest extends StateTransferFunctionalTest {
+
+   protected final boolean syncCommit;
+   protected final boolean writeSkew;
+   protected final boolean useSynchronization;
+
+   @Factory
+   @Override
+   public Object[] factory() {
+      return new Object[] {
+         new TotalOrderStateTransferFunctionalTest("dist-to-1pc-nbst", CacheMode.DIST_SYNC, true, false, false),
+         new TotalOrderStateTransferFunctionalTest("dist-to-2pc-nbst", CacheMode.DIST_SYNC, true, true, false),
+         new TotalOrderStateTransferFunctionalTest("repl-to-1pc-nbst", CacheMode.REPL_SYNC, true, true, true),
+         new TotalOrderStateTransferFunctionalTest("repl-to-2pc-nbst", CacheMode.REPL_SYNC, true, true, false),
+      };
+   }
+
+   public TotalOrderStateTransferFunctionalTest() {
+      this(null, null, false, false, false);
+   }
+
+   public TotalOrderStateTransferFunctionalTest(String cacheName, CacheMode mode, boolean syncCommit,
+                                                boolean writeSkew, boolean useSynchronization) {
+      super(cacheName);
+      this.cacheMode = mode;
+      this.syncCommit = syncCommit;
+      this.writeSkew = writeSkew;
+      this.useSynchronization = useSynchronization;
+   }
+
+   @Override
+   protected String parameters() {
+      return "{" + cacheName + "}";
+   }
+
+   protected void createCacheManagers() throws Throwable {
+      configurationBuilder = getDefaultClusteredCacheConfig(cacheMode, true);
+      configurationBuilder.transaction().transactionProtocol(TransactionProtocol.TOTAL_ORDER).syncCommitPhase(syncCommit)
+         .useSynchronization(useSynchronization)
+         .recovery().disable();
+      if (writeSkew) {
+         configurationBuilder.locking().isolationLevel(IsolationLevel.REPEATABLE_READ).writeSkewCheck(true);
+         configurationBuilder.versioning().enable().scheme(VersioningScheme.SIMPLE);
+      } else {
+         configurationBuilder.locking().isolationLevel(IsolationLevel.REPEATABLE_READ).writeSkewCheck(false);
+      }
+      configurationBuilder.clustering().stateTransfer().chunkSize(20);
+   }
+
+}
